+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to
   - [#4715](https://github.com/bpftrace/bpftrace/pull/4715)
 - Add `test` probes as means to improve runtime test latency
   - [#4753](https://github.com/bpftrace/bpftrace/pull/4753)
+- Add `long` and `ulong` int types
+  - [#4849](https://github.com/bpftrace/bpftrace/pull/4849)
 #### Changed
 - `uaddr` support PIE and dynamic library symbols.
   - [#4727](https://github.com/bpftrace/bpftrace/pull/4727)